requirements.txt
imageio
imageio-ffmpeg
opencv-python
pyserial
PySide6
pyside6-helpers
python-arduino-serial @ git@github.com:MrFrangipane/python-arduino-serial.git@main
